Add a States, CO and WhiteLabels folder to Assets and move all CO assets
