+++ src/Components/JeffcoComponents/JeffcoLandingPage/JeffcoLandingPage.tsx
@@ -1,7 +1,7 @@
 import { FormattedMessage } from 'react-intl';
 import { useNavigate } from 'react-router-dom';
 import { useConfig } from '../../Config/configHook';
-import jeffcolandingpage from '../../../Assets/COWhiteLabels/JeffcoAssets/jeffcolandingpage.png';
+import jeffcolandingpage from '../../../Assets/States/CO/WhiteLabels/JeffcoAssets/jeffcolandingpage.png';
 import Stack from '@mui/material/Stack';
 import { Typography, Button } from '@mui/material';
 import { useContext, useEffect } from 'react';
