+++ src/Components/Referrer/referrerDataInfo.tsx
@@ -1,17 +1,17 @@
 import MFBDEFAULT from '../../Assets/Logos/mfb_default_logo_header.png';
 import MFBCOLogo from '../../Assets/Logos/mfb_colorado_logo.png';
 import MFBNCLogo from '../../Assets/Logos/mfb_nc_logo.png';
-import BIAMFBLogo from '../../Assets/COWhiteLabels/BIA/biamfbcombinedlogo.png';
-import JHSAMFBLogo from '../../Assets/COWhiteLabels/JeffcoAssets/jeffcobrand.png';
-import VELogo from '../../Assets/COWhiteLabels/VillageExchange/villageExchangeLogo.png';
-import CCHMFBLogo from '../../Assets/COWhiteLabels/ColoradoCoalitionHomeless/cchcobrand.png';
-import LGSLogo from '../../Assets/COWhiteLabels/LetsGetSet/lgsLogo.png';
-import GACLogo from '../../Assets/COWhiteLabels/GetAheadColorado/gaclogo.png';
-import FIRCLogo from '../../Assets/COWhiteLabels/FircCobrand/FIRCLogo.png';
-import COMFBLogo from '../../Assets/colorado-mfb-logo.png';
-import DHSMFBLogo from '../../Assets/COWhiteLabels/DenverHumanServices/denverHSLogo.png';
-import CCIGLogo from '../../Assets/COWhiteLabels/CCIG/ccigLogo.png';
-import ECMFBLogo from '../../Assets/COWhiteLabels/EagleCounty/ECMFBLogo.png';
+import BIAMFBLogo from '../../Assets/States/CO/WhiteLabels/BIA/biamfbcombinedlogo.png';
+import JHSAMFBLogo from '../../Assets/States/CO/WhiteLabels/JeffcoAssets/jeffcobrand.png';
+import VELogo from '../../Assets/States/CO/WhiteLabels/VillageExchange/villageExchangeLogo.png';
+import CCHMFBLogo from '../../Assets/States/CO/WhiteLabels/ColoradoCoalitionHomeless/cchcobrand.png';
+import LGSLogo from '../../Assets/States/CO/WhiteLabels/LetsGetSet/lgsLogo.png';
+import GACLogo from '../../Assets/States/CO/WhiteLabels/GetAheadColorado/gaclogo.png';
+import FIRCLogo from '../../Assets/States/CO/WhiteLabels/FircCobrand/FIRCLogo.png';
+import COMFBLogo from '../../Assets/States/CO/mfb-logo.png';
+import DHSMFBLogo from '../../Assets/States/CO/WhiteLabels/DenverHumanServices/denverHSLogo.png';
+import CCIGLogo from '../../Assets/States/CO/WhiteLabels/CCIG/ccigLogo.png';
+import ECMFBLogo from '../../Assets/States/CO/WhiteLabels/EagleCounty/ECMFBLogo.png';
 
 const logoMap = {
   MFB_COLogo: MFBCOLogo,
