+++ src/Components/TwoOneOneComponents/TwoOneOneHeader/TwoOneOneHeader.js
@@ -6,8 +6,8 @@ import Box from '@mui/material/Box';
 import Paper from '@mui/material/Paper';
 import { useConfig } from '../../Config/configHook';
 import { Context } from '../../Wrapper/Wrapper';
-import twoOneOneMFBLogo from '../../../Assets/TwoOneOneAssets/twoOneOneMFBLogo.png';
-import twoOneOneLinks from '../../../Assets/TwoOneOneAssets/twoOneOneLinks';
+import twoOneOneMFBLogo from '../../../Assets/States/CO/WhiteLabels/TwoOneOneAssets/twoOneOneMFBLogo.png';
+import twoOneOneLinks from '../../../Assets/States/CO/WhiteLabels/TwoOneOneAssets/twoOneOneLinks';
 import LanguageIcon from '@mui/icons-material/Language';
 import ShareIcon from '@mui/icons-material/Share';
 import MenuIcon from '@mui/icons-material/Menu';
